backend/requirements.txt
Flask
Flask-Cors
python-dotenv
PlexAPI
requests  # For pyradarr/pysonarr, often a dependency
# For Sonarr/Radarr, we'll use direct requests as the libraries
# can be complex. This gives us more control.